Move get_bounds_var() to only place it's used in resample.py.
